+++ proguardprocessor/CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Unreleased
 
 - feat: emit processor version and type as attributes from all processors (#107) | @jairo-mendoza
-- feat: add telemetry support to proguard symbolicator
+- feat: handle stack frames that don't need symbolication (#106) | @jairo-mendoza
+- feat: add telemetry support to proguard symbolicator (#105) | @wolfgangcodes
 
 ## v0.0.4 [beta] - 2025/08/20